Check only nodes for affinity, which belong to the same profile
Previously the affinity plugin would check the whole cluster if a node
without an affinity pod. In turn this could block one maintenance
profile, if a node without a relevant pod is in maintenance-required
caused by an other profile. The affinity plugin now only considers
nodes, which are:
  1. In the maintenance-required state
  2. caused by the same profile.
The node affinity related stuff stays the same.